+++ ranking/tests/PageRank_Sim.cpp
@@ -1,10 +1,6 @@
 #include "PageRank.h"
-#include "data/Deserialize.h"
 #include "data/Document.h"
-#include "data/Gzip.h"
-#include "data/Reader.h"
 
-#include <filesystem>
 #include <fstream>
 #include <numeric>
 #include <unordered_map>
@@ -15,129 +11,56 @@ namespace {
 core::Config Config = core::Config("tests.conf");
 const std::string InputDirectory = std::string(Config.GetString("simulation_input_index_data_folder").Cstr());
 const std::string OutputFile = std::string(Config.GetString("pagerank_sim_out").Cstr());
-std::unordered_map<std::string, int> LinkToNode;
-std::unordered_map<int, std::string> NodeToLink;
-std::unordered_map<int, std::vector<int>> NodeConnections;
-int Nodes = 0;
+const bool WriteToFile = Config.GetInt("write_results_to_file") != 0;
 
-using namespace mithril;
-
-int GetLinkNode(const std::string& link) {
-    auto it = LinkToNode.find(link);
-    int nodeNo;
-    if (it == LinkToNode.end()) {
-        nodeNo = Nodes;
-        LinkToNode[link] = nodeNo;
-        NodeToLink[nodeNo] = link;
-        Nodes++;
-    } else {
-        nodeNo = it->second;
-    }
-
-    return nodeNo;
-}
-
-void Process() {
-    for (const auto& entry : std::filesystem::recursive_directory_iterator(InputDirectory)) {
-        if (!entry.is_regular_file()) {
-            continue;  // skip chunk dir
-        }
-
-        try {
-            data::Document doc;
-            {
-                data::FileReader file{entry.path().string().c_str()};
-                data::GzipReader gzip{file};
-                if (!data::DeserializeValue(doc, gzip)) {
-                    throw std::runtime_error("Failed to deserialize document: " + entry.path().string());
-                }
-            }
-
-            int fromNode = GetLinkNode(doc.url);
-            auto& vec = NodeConnections[fromNode];
-
-            for (const std::string& link : doc.forwardLinks) {
-                vec.push_back(GetLinkNode(link));
-            }
-        } catch (const std::exception& e) {
-            spdlog::error("\nError processing {}: {}", entry.path().string(), e.what());
-        }
-    }
-}
-}  // namespace
-
-int main(int /*argc*/, char* /*argv*/[]) {
-#if !defined(NDEBUG)
-    spdlog::set_level(spdlog::level::debug);
-#else
-    spdlog::set_level(spdlog::level::info);
-#endif
-
-    auto start = std::chrono::steady_clock::now();
-
-    spdlog::info("Starting page rank forward links simulation...");
-
-    Process();
-    const double tol = 1.0 / Nodes;
-
-    auto end = std::chrono::steady_clock::now();
-    auto processDuration = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
-
-    spdlog::info("Finished processing documents. Found {} links. Time taken: {} ms.", Nodes, processDuration);
-
-    spdlog::info("Building CSR Matrix with tolerance {:e}", tol);
-
-    start = std::chrono::steady_clock::now();
-
-    core::CSRMatrix m(Nodes);
-    std::vector<double> outDegree(Nodes, 0.0);
-
-    for (auto& [node, value] : NodeConnections) {
-        for (auto target : value) {
-            m.AddEdge(target, node, 1.0);
-        }
-
-        outDegree[node] = static_cast<double>(value.size());
-    }
-
-    m.Finalize();
-
-    for (int i = 0; i < m.values_.size(); ++i) {
-        if (outDegree[m.col_idx_[i]] > 0) {
-            m.values_[i] /= outDegree[m.col_idx_[i]];
-        }
-    }
-
-    end = std::chrono::steady_clock::now();
-    auto csrMatrixDuration = (std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count());
-
-    spdlog::info("Finished CSR matrix building process. Time taken: {} ms", csrMatrixDuration);
-    spdlog::info("Performing page rank....");
-
-    start = std::chrono::steady_clock::now();
-
-    PageRank algo(m, Nodes);
-
-    end = std::chrono::steady_clock::now();
-    auto duration = (std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count());
-    spdlog::info("Finished pagerank in: {} ms", duration);
-    spdlog::info("Total time taken: {} ms", (duration + csrMatrixDuration + processDuration));
+void WriteBackToFile() {
+    spdlog::info("Writing to human readable output file {}...", OutputFile);
 
     std::ofstream outFile;
     outFile.open(OutputFile);
 
-    std::vector<double> scores = algo.GetPageRanks();
-    std::vector<size_t> idx(scores.size());
+    std::vector<double>& scores = *mithril::pagerank::Results;
+    std::vector<size_t> idx(mithril::pagerank::DocumentCount);
     std::iota(idx.begin(), idx.end(), 0);
 
-    stable_sort(idx.begin(), idx.end(), [&scores](size_t i1, size_t i2) { return scores[i1] < scores[i2]; });
+    sort(idx.begin(), idx.end(), [&scores](size_t i1, size_t i2) {
+        auto it1 = mithril::pagerank::DocumentToNode->find(static_cast<mithril::data::docid_t>(i1));
+        auto it2 = mithril::pagerank::DocumentToNode->find(static_cast<mithril::data::docid_t>(i2));
+
+        double j1 = it1 == mithril::pagerank::DocumentToNode->end() ? 0 : scores[it1->second];
+        double j2 = it2 == mithril::pagerank::DocumentToNode->end() ? 0 : scores[it2->second];
+        return j1 < j2;
+    });
 
     for (size_t i = 0; i < idx.size(); ++i) {
-        outFile << NodeToLink[static_cast<int>(idx[i])] << ": " << scores[idx[i]] << std::endl;
+        size_t docID = idx[i];
+        auto it1 = mithril::pagerank::DocumentToNode->find(static_cast<mithril::data::docid_t>(docID));
+        if (it1 == mithril::pagerank::DocumentToNode->end()) {
+            outFile << "docid " << docID << " has no information\n";
+            continue;
+        }
+
+        int node = it1->second;
+
+        outFile << mithril::pagerank::ProcessLink((*mithril::pagerank::NodeToDocument)[node].url)
+                << " (docid: " << (*mithril::pagerank::NodeToDocument)[node].id
+                << "): " << (*mithril::pagerank::StandardizedResults)[node] << " (" << scores[node] << ")"
+                << std::endl;
     }
 
     outFile.close();
 
     spdlog::info("Finished writing to file: {}", OutputFile);
+}
+
+}  // namespace
+
+int main(int /*argc*/, char* /*argv*/[]) {
+    mithril::pagerank::PerformPageRank();
+
+    if (WriteToFile) {
+        WriteBackToFile();
+    }
+
     return 0;
 }