+++ CHANGELOG.md
@@ -54,6 +54,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - more metrics and attempts at improvements (#114)
 - old middle queue host cooldown combined with rate limit (#117)
 - introduced static ranking (#115)
+- add pagerank support for index (#118)
 
 ### Fixed
 - Fix copying body data from buffer (#24)